Alters project arch into hexagonal
