+++ source/Core/MongoDockerSample.Core.Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -1,0 +1,15 @@
+﻿using MongoDockerSample.Core.Application.Services;
+using MongoDockerSample.Core.Domain.Services;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    public static class ApplicationServiceCollectionExtensions
+    {
+        public static IServiceCollection AddServices(this IServiceCollection services)
+        {
+            services.AddScoped<IRegisterService , RegisterService>();
+
+            return services;
+        }
+    }
+}
