+++ source/Core/MongoDockerSample.Core.Application/MongoDockerSample.Core.Application.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.2.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MongoDockerSample.Core.Domain\MongoDockerSample.Core.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
