+++ source/Core/MongoDockerSample.Core.Domain/Models/Register.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace MongoDockerSample.Core.Domain.Models
+{
+    public class Register
+    {
+        public Guid Key { get; set; }
+
+        public string Value { get; set; }
+    }
+}
