--- source/Domain/MongoDocker.Sample.Domain.Contract/DTO/MongoDbConfigurationValues.cs
@@ -1,9 +1,0 @@
-﻿namespace MongoDocker.Sample.Domain.Contract.DTO
-{
-    public class MongoDbConfigurationValues
-    {
-        public string Server { get; set; }
-        public string Database { get; set; }
-        public string Collection { get; set; }
-    }
-}
