--- source/Domain/MongoDocker.Sample.Domain.Contract/Exception/Base/CustomError.cs
@@ -1,16 +1,0 @@
-﻿using System.Net;
-
-namespace MongoDocker.Sample.Domain.Contract.Exception.Base
-{
-    public abstract class CustomError
-    {
-        public HttpStatusCode StatusCode { get; set; }
-        public string Message { get; }
-
-        public CustomError(HttpStatusCode statusCode, string error)
-        {
-            StatusCode = statusCode;
-            Message = error;
-        }
-    }
-}
