--- source/Domain/MongoDocker.Sample.Domain.Contract/MongoDocker.Sample.Domain.Contract.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<doc>
-    <assembly>
-        <name>MongoDocker.Sample.Domain.Contract</name>
-    </assembly>
-    <members>
-    </members>
-</doc>
