--- source/Domain/MongoDocker.Sample.Domain.Service/MongoDocker.Sample.Domain.Service.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MongoDocker.Sample.Domain.Contract\MongoDocker.Sample.Domain.Contract.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="MongoDocker.Sample.Domain.Service.xml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
