--- source/Domain/MongoDocker.Sample.Domain.Service/MongoDocker.Sample.Domain.Service.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0"?>
-<doc>
-    <assembly>
-        <name>MongoDocker.Sample.Domain.Service</name>
-    </assembly>
-    <members>
-        <member name="T:MongoDocker.Sample.Domain.Service.Interfaces.IMongoDbService">
-            <summary>
-            Service responsible for create, request, update and delete a value
-            </summary>
-        </member>
-        <member name="M:MongoDocker.Sample.Domain.Service.Interfaces.IMongoDbService.InsertValueAsync(System.String)">
-            <summary>
-            Inserts new register
-            </summary>
-            <param name="value">String value for the new register</param>
-            <returns>New register's identifier</returns>
-        </member>
-        <member name="M:MongoDocker.Sample.Domain.Service.Interfaces.IMongoDbService.UpdateValueAsync(System.Guid,System.String)">
-            <summary>
-            Update some register's value
-            </summary>
-            <param name="key">StRegister's identifier</param>
-            <param name="newValue">Register's new value</param>
-        </member>
-        <member name="M:MongoDocker.Sample.Domain.Service.Interfaces.IMongoDbService.DeleteValueAsync(System.Guid)">
-            <summary>
-            Removes some register from collection
-            </summary>
-            <param name="key">StRegister's identifier</param>
-        </member>
-        <member name="M:MongoDocker.Sample.Domain.Service.Interfaces.IMongoDbService.GetValueAsync(System.Guid)">
-            <summary>
-            Gets register value
-            </summary>
-            <param name="key">Register's key</param>
-            <returns>Registers object <see cref="T:MongoDocker.Sample.Domain.Contract.DTO.MongoDbRegister"/></returns>
-        </member>
-        <member name="M:MongoDocker.Sample.Domain.Service.Interfaces.IMongoDbService.GetValuesAsync">
-            <summary>
-            Gets all register values
-            </summary>
-            <returns>Registers list <see cref="T:MongoDocker.Sample.Domain.Contract.DTO.MongoDbRegister"/></returns>
-        </member>
-    </members>
-</doc>
