--- source/Infrastructure/MongoDocker.Sample.Infrastructure.Log/Dockerfile
@@ -1,9 +1,0 @@
-FROM datalust/seqcli:5.1.251
-
-RUN apt-get update && apt-get install -y curl
-
-COPY ./docker-entrypoint.sh /
-
-RUN ["chmod", "+x", "/docker-entrypoint.sh"]
-
-ENTRYPOINT ["/docker-entrypoint.sh"]
