--- source/Infrastructure/MongoDocker.Sample.Infrastructure.Provider/MongoDocker.Sample.Infrastructure.Provider.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="MongoDB.Bson" Version="2.10.1" />
-    <PackageReference Include="MongoDB.Driver" Version="2.10.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Domain\MongoDocker.Sample.Domain.Contract\MongoDocker.Sample.Domain.Contract.csproj" />
-    <ProjectReference Include="..\..\Domain\MongoDocker.Sample.Domain.Service\MongoDocker.Sample.Domain.Service.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="MongoDocker.Sample.Infrastructure.Provider.xml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
