--- source/Infrastructure/MongoDocker.Sample.Infrastructure.Provider/MongoDocker.Sample.Infrastructure.Provider.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<doc>
-    <assembly>
-        <name>MongoDocker.Sample.Infrastructure.Provider</name>
-    </assembly>
-    <members>
-        <member name="M:MongoDocker.Sample.Infrastructure.Provider.MongoDbService.#ctor(MongoDocker.Sample.Domain.Contract.DTO.MongoDbConfigurationValues)">
-            <summary>
-            Receives MongoDb configuration values through dependency injection
-            </summary>
-            <param name="mongoDbConfigurationValues"></param>
-        </member>
-    </members>
-</doc>
