+++ source/Infrastructure/MongoDockerSample.Infrastructure.Repository/DependencyInjection/RepositoryServiceCollectionExtensions.cs
@@ -1,0 +1,29 @@
+﻿using AutoMapper;
+using MongoDockerSample.Core.Domain.Repositories;
+using MongoDockerSample.Infrastructure.Repository;
+using MongoDockerSample.Infrastructure.Repository.Repositories;
+using System;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    public static class RepositoryServiceCollectionExtensions
+    {
+        public static IServiceCollection AddRepository(
+            this IServiceCollection services,
+            RepositoryConfiguration configuration)
+        {
+            if (configuration is null)
+            {
+                throw new ArgumentNullException(nameof(configuration));
+            }
+
+            services.AddSingleton(configuration);
+
+            services.AddAutoMapper(typeof(RepositoryMapperProfile));
+            
+            services.AddScoped<IRegisterRepository , RegisterRepository>();
+
+            return services;
+        }
+    }
+}
