--- source/MongoDocker.Sample.sln
@@ -1,55 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29609.76
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Ui", "Ui", "{997C2735-364B-44E2-8AD7-E406E9AE4AB3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDocker.Sample.Ui.Api", "Ui\MongoDocker.Sample.Ui.Api\MongoDocker.Sample.Ui.Api.csproj", "{42FEBAB1-399A-4F83-8516-72F67E71850A}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{F90214D3-5E26-43BA-AD19-818730948211}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain", "Domain", "{8B248B19-6820-448E-A63E-4B4523AA7AE9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDocker.Sample.Domain.Service", "Domain\MongoDocker.Sample.Domain.Service\MongoDocker.Sample.Domain.Service.csproj", "{58C546D0-4854-44FB-A073-4B39240D59A4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDocker.Sample.Infrastructure.Provider", "Infrastructure\MongoDocker.Sample.Infrastructure.Provider\MongoDocker.Sample.Infrastructure.Provider.csproj", "{29A3B287-6C5C-40F2-8973-00D4E2EB01CC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDocker.Sample.Domain.Contract", "Domain\MongoDocker.Sample.Domain.Contract\MongoDocker.Sample.Domain.Contract.csproj", "{2F4DDC95-2735-4E6F-AAD8-3F6D87277B63}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{42FEBAB1-399A-4F83-8516-72F67E71850A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{42FEBAB1-399A-4F83-8516-72F67E71850A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{42FEBAB1-399A-4F83-8516-72F67E71850A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{42FEBAB1-399A-4F83-8516-72F67E71850A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{58C546D0-4854-44FB-A073-4B39240D59A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{58C546D0-4854-44FB-A073-4B39240D59A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{58C546D0-4854-44FB-A073-4B39240D59A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{58C546D0-4854-44FB-A073-4B39240D59A4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{29A3B287-6C5C-40F2-8973-00D4E2EB01CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29A3B287-6C5C-40F2-8973-00D4E2EB01CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{29A3B287-6C5C-40F2-8973-00D4E2EB01CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{29A3B287-6C5C-40F2-8973-00D4E2EB01CC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2F4DDC95-2735-4E6F-AAD8-3F6D87277B63}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2F4DDC95-2735-4E6F-AAD8-3F6D87277B63}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2F4DDC95-2735-4E6F-AAD8-3F6D87277B63}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2F4DDC95-2735-4E6F-AAD8-3F6D87277B63}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{42FEBAB1-399A-4F83-8516-72F67E71850A} = {997C2735-364B-44E2-8AD7-E406E9AE4AB3}
-		{58C546D0-4854-44FB-A073-4B39240D59A4} = {8B248B19-6820-448E-A63E-4B4523AA7AE9}
-		{29A3B287-6C5C-40F2-8973-00D4E2EB01CC} = {F90214D3-5E26-43BA-AD19-818730948211}
-		{2F4DDC95-2735-4E6F-AAD8-3F6D87277B63} = {8B248B19-6820-448E-A63E-4B4523AA7AE9}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {715F75F5-0494-4273-881B-845E18D9D0BC}
-	EndGlobalSection
-EndGlobal
