--- source/Ui/MongoDocker.Sample.Ui.Api/MongoDocker.Sample.Ui.Api.csproj
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNet.WebApi.Core" Version="5.2.7" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Seq.Extensions.Logging" Version="4.0.2" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
-    <PackageReference Include="Swashbuckle.AspNetCore.Swagger" Version="4.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Domain\MongoDocker.Sample.Domain.Service\MongoDocker.Sample.Domain.Service.csproj" />
-    <ProjectReference Include="..\..\Infrastructure\MongoDocker.Sample.Infrastructure.Provider\MongoDocker.Sample.Infrastructure.Provider.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="MongoDocker.Sample.Ui.Api.xml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
