--- source/Ui/MongoDocker.Sample.Ui.Api/MongoDocker.Sample.Ui.Api.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0"?>
-<doc>
-    <assembly>
-        <name>MongoDocker.Sample.Ui.Api</name>
-    </assembly>
-    <members>
-        <member name="T:MongoDocker.Sample.Ui.Api.Controllers.ValuesController">
-            <summary>
-            Controller responsible for create, request, update and delete a value
-            </summary>
-        </member>
-        <member name="M:MongoDocker.Sample.Ui.Api.Controllers.ValuesController.#ctor(MongoDocker.Sample.Domain.Service.Interfaces.IMongoDbService)">
-            <summary>
-            Receives IMongoDbService instance from dependency injection
-            </summary>
-            <param name="mongoDbService"></param>
-        </member>
-        <member name="M:MongoDocker.Sample.Ui.Api.Controllers.ValuesController.GetAsync(System.Guid)">
-            <summary>
-            Returns full object base on its key.
-            </summary>
-            <param name="key">Object key</param>
-            <returns>MongoDbRegistrer <see cref="T:MongoDocker.Sample.Domain.Contract.DTO.MongoDbRegister"/></returns>
-        </member>
-        <member name="M:MongoDocker.Sample.Ui.Api.Controllers.ValuesController.PostAsync(System.String)">
-            <summary>
-            Inserts new object.
-            </summary>
-            <param name="value">Object value</param>
-            <returns>Object identifier <see cref="T:System.Guid"/> </returns>
-        </member>
-        <member name="M:MongoDocker.Sample.Ui.Api.Controllers.ValuesController.GetAllAsync">
-            <summary>
-            Returns a list of objects
-            </summary>
-            <returns>MongoDbRegistrer <see cref="T:MongoDocker.Sample.Domain.Contract.DTO.MongoDbRegister"/></returns>
-        </member>
-        <member name="M:MongoDocker.Sample.Ui.Api.Controllers.ValuesController.PutAsync(System.Guid,System.String)">
-            <summary>
-            Updates object value.
-            </summary>
-            <param name="key">Objects key</param>
-            <param name="value">Objects new value</param>
-        </member>
-        <member name="M:MongoDocker.Sample.Ui.Api.Controllers.ValuesController.DeleteAsync(System.Guid)">
-            <summary>
-            Deletes object.
-            </summary>
-            <param name="key">Objects key</param>
-        </member>
-    </members>
-</doc>
