--- source/Ui/MongoDocker.Sample.Ui.Api/appsettings.json
@@ -1,20 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "MongoDb": {
-    "Server": "mongodb://mongo",
-    "Database": "local",
-    "Collection": "default"
-  },
-  "LogSeq": {
-    "ServerUrl": "http://logseq",
-    "MinimumLevel": "Trace",
-    "LevelOverride": {
-      "Microsoft": "Warning"
-    }
-  }
-}
