+++ source/Ui/MongoDockerSample.Ui.Api/MongoDockerSample.Ui.Api.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="Seq.Extensions.Logging" Version="4.0.2" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore.Swagger" Version="4.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Core\MongoDockerSample.Core.Application\MongoDockerSample.Core.Application.csproj" />
+    <ProjectReference Include="..\..\Core\MongoDockerSample.Core.Domain\MongoDockerSample.Core.Domain.csproj" />
+    <ProjectReference Include="..\..\Infrastructure\MongoDockerSample.Infrastructure.Repository\MongoDockerSample.Infrastructure.Repository.csproj" />
+  </ItemGroup>
+
+</Project>
