+++ source/Ui/MongoDockerSample.Ui.Api/appsettings.Development.json
