+++ source/Ui/MongoDockerSample.Ui.Api/appsettings.json
@@ -1,0 +1,20 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "RepositoryConfiguration": {
+    "MongoServer": "mongodb://mongo",
+    "MongoDatabase": "local",
+    "MongoCollection": "default"
+  },
+  "LogSeq": {
+    "ServerUrl": "http://logseq",
+    "MinimumLevel": "Trace",
+    "LevelOverride": {
+      "Microsoft": "Warning"
+    }
+  }
+}
